requirements.txt
pydantic~=1.9
requests~=2.28
pytest~=7.0.1
allure-pytest
